Remove obsolete handler classes
