--- addon/editor/input-handlers/copy-handler.ts
@@ -1,10 +1,0 @@
-import CutCopyHandler from '@lblod/ember-rdfa-editor/editor/input-handlers/cut-copy-handler';
-
-/**
- * @module contenteditable-editor
- * @class CopyHandler
- * @constructor
- */
-export default class CopyHandler extends CutCopyHandler {
-  deleteSelection = false;
-}
