--- addon/editor/input-handlers/cut-handler.ts
@@ -1,10 +1,0 @@
-import CutCopyHandler from '@lblod/ember-rdfa-editor/editor/input-handlers/cut-copy-handler';
-
-/**
- * @module contenteditable-editor
- * @class CutHandler
- * @constructor
- */
-export default class CutHandler extends CutCopyHandler {
-  deleteSelection = true;
-}
