--- addon/editor/input-handlers/escape-handler.ts
@@ -1,29 +1,0 @@
-import { InputHandler } from './input-handler';
-import RawEditor from '@lblod/ember-rdfa-editor/utils/ce/raw-editor';
-import { isKeyDownEvent } from '@lblod/ember-rdfa-editor/editor/input-handlers/event-helpers';
-
-/**
- * EscapeHandler, an event handler to handle escape.
- *
- * @module contenteditable-editor
- * @class EscapeHandler
- * @constructor
- */
-export default class EscapeHandler extends InputHandler {
-  constructor({ rawEditor }: { rawEditor: RawEditor }) {
-    super(rawEditor);
-  }
-
-  isHandlerFor(event: Event) {
-    return isKeyDownEvent(event) && event.key === 'Escape';
-  }
-
-  handleEvent(_: KeyboardEvent) {
-    const activeElement = document.activeElement;
-    if (activeElement) {
-      (activeElement as HTMLElement).blur();
-    }
-
-    return { allowPropagation: false };
-  }
-}
